+++ test/integration/062_defer_state_test/changed_models_bad/table_model.sql
@@ -1,0 +1,2 @@
+{{ config(materialized='table') }}
+select * from {{ ref('ephemeral_model') }}